requirements.txt
flask
flask-cors
pandas
plotly
requests
python-dotenv
pytest
openai
psycopg2-binary
slowapi
